+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## unreleased
 
 * Updated `nalgebra` to 0.10.
+* Added `Solve` trait for solving a linear matrix equation.
+* Added `Inverse` for computing the multiplicative inverse of a matrix.